# Exclude edge cases where get_relative_url depends on CWD