--- a/mkdocs/utils/__init__.py
+++ b/mkdocs/utils/__init__.py
@@ -259,7 +259,7 @@ def get_relative_url(url, other):
         # Remove filename from other url if it has one.
         parts = posixpath.split(other)
         other = parts[0] if '.' in parts[1] else other
-    relurl = posixpath.relpath(url, other)
+    relurl = posixpath.relpath('/' + url, '/' + other)
     return relurl + '/' if url.endswith('/') else relurl