When only one of the two passed paths starts with a slash, this function would produce results that depend on (and expose parts of) the actual current working directory.
In a similar manner, it was also possible to "break out" of the "current directory" by starting one of the paths with `../../..` etc.

Additionally, the fact that paths always end up being resolved relative to the current working directory made this function less efficient than it needs to be.
Fun fact: `get_relative_url('path_a', 'path_b')` actually ended up looking up how to get to `/current/working/directory/path_a` from `/current/working/directory/path_b`.

Luckily, none of these behaviors can ever actually come into effect (the function is always called without a leading slash and without paths deliberately trying to escape), but the fix is still good to reduce surprise.

*The actual fix* is that we make the leading slash ignored (or quite the opposite - we always add it).
Now when either of the two arguments try to go up higher than the top level, they just end up at the top level (e.g. `foo/../..` is effectively `.`), only then the "relative" calculation happens.

---

This pull request makes the behavior 100% the same as in https://github.com/mkdocs/mkdocs/pull/2272, but without the disadvantage of being a "custom replacement for the standard lib function". It also brings 35% of the performance advantage.

Build times of a site that has ~350 pages:

* When running from directory `/home/oprypin/repos/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/athena-website`:
  * Before: 11.1 sec
  * After: 8.22 sec
* When running from directory `/home/oprypin/repos/athena-website`:
  * Before: 9.02 sec
  * After: 8.15 sec

(meanwhile, https://github.com/mkdocs/mkdocs/pull/2272 achieves 6.65 sec)