@oprypin 2021-02-06
How can you keep saying that the standard lib is correct, if **the usage of it is outright wrong**. A function that resolves paths **relative to the current working directory** is being used to resolve abstract paths. How is it in any way OK that `get_relative_url` produces a different result depending on which directory you happen to be running in?

To clarify: this comment describes the current state, not the state after this pull request closes those loopholes.

I will print out an expanded list here, sure.